feat: support new database driver `sqlite` (#49)
This will likely not work for large db files that cannot be opening
in memory. For now, the native binding approach wasn't working either,
so this will have to do. Also, it's nice that this provides a path for
compatibility in the browser-only context.

Switch to `esbuild` for bundling, remove unused dev deps.